% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenGovFunctions.R
\name{getQueryWaterQuality}
\alias{getQueryWaterQuality}
\title{Get Water Quality Query}
\usage{
getQueryWaterQuality(sql, con = NULL)
}
\arguments{
\item{sql}{character string of the query}

\item{con}{supply to reuse a sql connection}
}
\description{
get the results of a sql query to the water quality database
}
