% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenGovFunctions.R
\name{getWQData}
\alias{getWQData}
\title{Get Water Quality DAta}
\usage{
getWQData(lwrBnd_date = "2015-1-1", uprBnd_date = "2016-4-12",
  date_format = "\%Y-\%m-\%d", chemical = NULL, PWS_ID = NULL)
}
\arguments{
\item{lwrBnd_date}{- character string of date}

\item{uprBnd_date}{- character string of date}

\item{date_format}{- the format of the given dates}

\item{chemical}{- specify a chemical to select for}

\item{PWS_ID}{- specify a water-supplier ID to select for}
}
\description{
query water quality data by dates
chemical, and or PWS_ID. Dates are
converted to julian.
}

