% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenGovFunctions.R
\name{gmapsQuery}
\alias{gmapsQuery}
\title{Google Maps Query}
\usage{
gmapsQuery(address)
}
\arguments{
\item{address}{character string of street address to geocode}
}
\description{
geocode an address with google api (2500 per day, 10 per sec)
}

