% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenGovFunctions.R
\name{loadSupplierSummaryData}
\alias{loadSupplierSummaryData}
\title{Load Supplier Summary Data}
\usage{
loadSupplierSummaryData(reload = NULL, env = globalenv(), yield_top = 5,
  dataDirectory = paste0("~/Documents/", "greengov-challenge", "/Data/"))
}
\arguments{
\item{reload}{- y/n/null should data be reconstructed}

\item{env}{- environment to load data into}

\item{yield_top}{numeric - number of records to deliver for each supplier}

\item{dataDirectory}{- character string of where the data is located}
}
\description{
load a summary of waste water violations and chemicals present
for each supplier
}
