% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenGovFunctions.R
\name{loadSupplierWQData}
\alias{loadSupplierWQData}
\title{Load Supplier Water Quality Data}
\usage{
loadSupplierWQData(reload = NULL, env = globalenv(),
  dataDirectory = "~/Documents/greengov-challenge/Data/")
}
\arguments{
\item{reload}{- y/n/null should data be reconstructed}

\item{env}{- environment to load data into}

\item{dataDirectory}{- character string of where the data is located}
}
\description{
load the supplier-oriented water quality data
}
