% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenGovFunctions.R
\name{loadWasteWaterViolationsData}
\alias{loadWasteWaterViolationsData}
\title{Load Waste Water Violations}
\usage{
loadWasteWaterViolationsData(reload = NULL, env = globalenv(),
  dataDirectory = "~/Documents/greengov-challenge/Data/")
}
\arguments{
\item{reload}{character ("y"/"n") - should the data be re-constructed?}

\item{env}{environment to load the data into}

\item{dataDirectory}{character - where is the raw data located?}
}
\description{
get the waste water violations data
}

