% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenGovFunctions.R
\name{sendQueryWaterQuality}
\alias{sendQueryWaterQuality}
\title{Send Water Quality Query}
\usage{
sendQueryWaterQuality(sql, con = NULL)
}
\arguments{
\item{sql}{character string of the query}

\item{con}{supply to reuse a sql connection}
}
\description{
send a sql query to the waterQuality sqlite3 database
}

